move adjustPackageJson to own file
