+++ src/verdaccio/publish-package.ts
@@ -1,69 +1,13 @@
-import type { PublishPackagesAndInstallArgs } from '.'
-import type { PackageJson } from '../types'
+import type { PublishPackageArgs } from '.'
 import path from 'node:path'
 import process from 'node:process'
-import destr from 'destr'
-import fs from 'fs-extra'
 import { CLI_NAME } from '../constants'
+import { adjustPackageJson } from '../utils/adjust-package-json'
 import { createTempNpmRc } from '../utils/create-temp-npm-rc'
 import { getSourcePackageJsonPath } from '../utils/file'
 import { logger } from '../utils/logger'
 import { promisifiedSpawn } from '../utils/promisified-spawn'
-import { registerCleanupTask } from './cleanup-tasks'
 import { REGISTRY_URL } from './verdaccio-config'
-
-type AdjustPackageJsonArgs = Omit<PublishPackageArgs, 'source'> & {
-  sourcePkgJsonPath: string
-}
-
-/**
- * Edit package.json to:
- * - Adjust version to temporary "secco" version
- * - Change version selectors for dependencies that will be published (to make sure they are installed)
- */
-function adjustPackageJson({ sourcePkgJsonPath, packageName, packageNamesToFilePath, packagesToPublish, ignorePackageJsonChanges, versionPostfix }: AdjustPackageJsonArgs) {
-  // Check if package depends on any other package that will be published. Adjust version selector to point to "secco" version so that local registry is used for those dependencies.
-
-  const sourcePkgJsonString = fs.readFileSync(sourcePkgJsonPath, 'utf-8')
-  const sourcePkgJson = destr<PackageJson>(sourcePkgJsonString)
-
-  // Overwrite version with "secco" name
-  sourcePkgJson.version = `${sourcePkgJson.version}-${CLI_NAME}-${versionPostfix}`
-
-  packagesToPublish.forEach((pkgToPublish) => {
-    if (sourcePkgJson.dependencies?.[pkgToPublish]) {
-      const srcPath = getSourcePackageJsonPath(pkgToPublish, packageNamesToFilePath)
-
-      if (!srcPath)
-        return
-
-      const currentVersion = destr<PackageJson>(fs.readFileSync(srcPath, 'utf-8'))?.version
-
-      if (currentVersion)
-        sourcePkgJson.dependencies[pkgToPublish] = `${currentVersion}-${CLI_NAME}-${versionPostfix}`
-    }
-  })
-
-  const tempSourcePkgJsonString = JSON.stringify(sourcePkgJson)
-
-  const revertIgnorePackageJsonChanges = ignorePackageJsonChanges(packageName, [sourcePkgJsonString, tempSourcePkgJsonString])
-
-  fs.outputFileSync(sourcePkgJsonPath, tempSourcePkgJsonString)
-
-  return {
-    newPackageVersion: sourcePkgJson.version,
-    revertAdjustPackageJson: registerCleanupTask(() => {
-      // Restore original package.json file
-      fs.outputFileSync(sourcePkgJsonPath, sourcePkgJsonString)
-      revertIgnorePackageJsonChanges()
-    }),
-  }
-}
-
-type PublishPackageArgs = Omit<PublishPackagesAndInstallArgs, 'destination'> & {
-  packageName: string
-  versionPostfix: string
-}
 
 export async function publishPackage({ packageName, packagesToPublish, packageNamesToFilePath, versionPostfix, source, ignorePackageJsonChanges }: PublishPackageArgs) {
   const sourcePkgJsonPath = getSourcePackageJsonPath(packageName, packageNamesToFilePath)
